assignment-pub-sub/pyproject.toml
[project]
name = "assignment-pub-sub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-pubsub>=2.0.0",
]

[project.scripts]
producer = "producer.producer:main"
consumer = "consumer.consumer:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["producer*", "consumer*"]

[tool.uv]
package = true
